[IMP] mail_enterprise: remove createMessagingMenu during tests
The createMessagingMenu helper was used during tests to mount
a messaging menu and test it. However, this approach is not
very realistic and blocks some waiting PRs. In order to get closer
from the reality, let's instanciate a webClient and a systrayService
for the messagingMenu to be present at all time.

closes odoo/enterprise#28747

Related: odoo/odoo#94301
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>