[MOV] website_appointment: move static option file
With community counter-part at [1], the website UI is moved to the
backend. This commit prepares the work by moving a file (without
modifying it) to the destination where it will be refactored. This is
done to preserve history and ease forward-ports.

[1]: https://github.com/odoo/odoo/pull/89223

See merge commit for more information.

task-2687506

Part-of: odoo/enterprise#28448