[FIX] l10n_be_hr_payroll: Fix eco-voucher generation
This condition was added to exclude employees who already received their
eco-vouchers before june (aka employees who left the company and received
the amount at that time).

But the employees who stayed in the company were excluded too as we were
including june, which the month where we actually pay the eco-vouchers.

closes odoo/enterprise#28871

X-original-commit: 66322b6dbe14dbe47cc5bd7a018d50e1115bc26b
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>