[REF] documents_spreadsheet: extract edition features in a new module
This commit is a step toward dashboard introduction.
In order to allow the edition of the future dashboards without dependency
to `documents`, we have to extract the edition features from
`documents_spreadsheet` (pivot/list/graph insertion, side panels, ...)

So, a new module is created: `spreadsheet_edition` which contains all
theses features. `documents_spreadsheet` depends now on this module. The
plan for dashboards is to make the dashboard module dependent of this one.

Part of task-id 2799682

Part-of: odoo/enterprise#29105
Co-authored-by: Pierre Rousseau <pro@odoo.com>
Co-authored-by: Lucas Lefèvre <lul@odoo.com>