[FIX] account_intrastat: always use unknown vat
In the case of intrastat, there is a specific vat code used when the vat
code is unknown. This code is QV999999999999. For some countries, there
is a different vat code that is also used when the vat code is not known
and the partner is an individual (rather than a company).

At present the 'unknown' code is not being used, except when the
country_id is not set on the partner.

This commit generalises the use of the unknown vat code so that, if the
vat on the partner is not set, the unknown vat code is used. When the
partner is not a company, the same code is used, unless the company's
country code is present in a newly defined tuple.

If the partner's country code is in this tuple, then the 'QN999999999999'
unknown individual vat code is used, since currently the unknown
individual is'QN999999999999' for all known countries that make this
distinction.

closes odoo/enterprise#29606

Task-id: 2894443
X-original-commit: 9728aec373877887789a7097ec924088842ccea7
Signed-off-by: Brice Bartoletti <bib@odoo.com>