[IMP] spreadsheet, documents_spreadsheet: odoo chart support
This commit introduce the support of Odoo charts (line, bar and pie)
which retrieve the data from the server instead of the cells of the
spreadsheet.

For that, a new data source has been created to handle the fetching
of the chart data.

To insert a new chart, a new button has been added in the control panel
of the graph views.

Task-id 2822603

closes odoo/enterprise#28894

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>