[FIX] l10n_uk_reports: received date of obligations correctly parsed
How to reproduce the bug ?

- Install base,account_accountant and l10n_uk_reports
- In Settings, in Users & Companies > Companies, select the UK company
and set a VAT number.
- In Accounting, create an one-year old invoice or with a date of the
previous year.
- Then, go to Reporting > Tax Report and click on Connect To HMRC
- Then, go to Reporting > (United Kingdom) VAT Obligations.
- The Received Submission date field is not set.

What is the bug ?

When a customer send the HMRC VAT obligations, a response is sent. In
the response, there might be a field related to the reception date.
Currently, we are excepting that this field has 'received_date' as key.
However, according to the documentation, the key of the field is
'received'. Because of this, in Odoo, the field is always blank.
The URL of the server used in DEBUG mode was also deprecated.

opw-2842878

closes odoo/enterprise#30395

X-original-commit: f653b0f9fd6a303ac3691b0622b0b4e461c84c8f
Signed-off-by: Adrien Minet <admi@odoo.com>
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Minet Adrien (admi) <admi@odoo.com>