[ADD] website_delivery_fedex_locations: add module
This commit gives access to pick up delivery to the website clients.

When enabled, the "Fedex Location" gives the possibility to choose dropoff
locations close to the client's address.

Since the Location API is really close to the Rate and Ship APIs, the
FedexLocationsRequest class is a child of the FedexRequest.

To be able to add custom the chosen location to the 'ship', 'rate' and
'return' srm requests, the _fedex_update_srm function was added to the
delivery carrier. This function takes the picking or order and the srm
object, so that we have access to all current information.

closes odoo/enterprise#25292

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>