[FIX] web_dashboard: missing string attr in fake fields
The rendering of a Widget tag in the dashboard view requires the
creation of a legagy record datapoint when the widget to instantiate is
legagy. It turns out that the attr fields on the record was badly built:
the fake fields (linked to aggretate/formula) didn't have a string attr.
For that reason, some dashboard views with a pie chart widget did crash
if the measure used was the name of a fake field.

This commit assures that every field has a sting attr.

Part-of: odoo/enterprise#27879