[FIX] account_bank_statement_import_camt: correct amount in currency
In case of a batch transaction, and if no total amount is provided, we
cannot rely on the entry amount to compare with converted transaction
amount. We should retrieve the entry details amount first and fallback
on the entry amount. If not detailed amount is provided in the journal
currency, then we should fallback on the generic return, but with adding
the rate too to avoid returning a non-converted amount.

closes odoo/enterprise#29056

X-original-commit: a50d8e396795f7a77900208018d5d5a16e6c63b1
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>