[REF] documents_spreadsheet: decouple abstract action from documents
This commit is a step to split the `documents_spreadsheet` module in two.
All spreadsheet related code and models will be moved to a new module which
won't depend on the `documents` module.

In this commit we remove the dependency between the abstract spreadsheet
action and the `documents.document` model and templates.

Part-of: odoo/enterprise#28950