[FIX] account_reports: fix tour failing randomly
Before this commit, there was a race condition with the tour
"account_reports_widgets":
 - it toggles a first dropdown and selects an item which triggers
   a re-rendering
 - it toggles a second dropdown
 - the re-rendering ends and there is a repaint, so the second
   dropdown is folded again
 - it tries to select an item in the second dropdown, and can't
   find it.

This commit fixes the issue by waiting for the first dropdown to
be closed (i.e. waiting the re-rendering) before going further.
Note that this is a real issue with those custom dropdown items in
the account client action, but not easy to reproduce on a real
environment.

Part-of: odoo/enterprise#27879