[FIX] account_asset: wrong smart buttons on bills
- Create a bill that creates a deferred revenue and an asset
- There is only one smart button with, as name, the first of the deferred revenue or asset.

The problem is that the view takes the type of the first
asset/deferred revenue/deferred expenses in the move.

We have to create a button for each type.
We can remove the field `linked_asset_type` as it is now useless.
We renamed some fields to make their role more obvious.

task-2894690

closes odoo/enterprise#28926

Related: odoo/upgrade#3634
Signed-off-by: William André (wan) <wan@odoo.com>