[FIX] documents_spreadsheet: fix download button in list view
Downloading odoo spreadsheet in list view gives unusable binary content instead of .xlsx file
as in kanban view. It happens because spreadsheet is detected by value of
`handler` field, which is not available in tree view.

Fix it by adding invisible field `handler` to the spreadsheet tree view.

STEPS:
* Create Odoo Spreadsheet
* Switch to list view, select the file and click download button

opw-2888594

closes odoo/enterprise#29234

X-original-commit: d6e0d7283b4534c2e5a4ea73012ef8f22b13e786
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>