[REF] documents_spreadsheet: decouple spreadsheet dialog from documents
This commit is a step to split the `documents_spreadsheet` module in two.
All spreadsheet related code and models will be moved to a new module which
won't depend on the `documents` module.

In this commit we remove the dependency between the spreadsheet selector
dialog and `documents.documents`.

A new panel will be introduced very soon to link pivots/lists/etc to
another model then `document.document`

Part-of: odoo/enterprise#28950
Co-authored-by: Pierre Rousseau <pro@odoo.com>