[IMP] timesheet_grid, sale_timesheet_enterprise: add overtime info to the My Timesheet grid view
This commit adds overtime information to the grid view in the "My Timesheet"
page of the Timesheets app.
It adds the daily and total overtime in the total row of the grid.
The overtime can be positive or negative. When the overtime is zero, it is
not printed.

The improvement that this task introduces ensures that all employees can
easily keep track what they have yet to timesheet as well as whether they
have been productive enough at a glance, from their defualt view in the
Timesheet app.

task-2878635

closes odoo/enterprise#29023

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>