[FIX] l10n_be_reports: 281.50 dealing with legacy data
The aim of this commit is to take into account the legacy data of Odoo in order
to have correct results reported.

Context:
Odoo set a partner_id on the account.move.line.
Currently, the partner is always the commercial_partner_id A.K.A. the
"ultimate parent partner".
In the past, that hasn't always been the case and it was possible that Odoo
sets the partner directly.

Previous to this commit:
281.50 feature only take care of the parent and doesn't consider the account
move line on which the children has been set and thus doesn't even report
those amount.
There isn't any form 281.50 created for the child and the amounts relative to
this child are "lost".

After this commit:
281.50 feature takes those account.move.line into account.
All the amounts are gathered on the parent partner.

closes odoo/enterprise#28972

Pr: https://github.com/odoo/enterprise/pull/28065
Task-id: 2871601
X-original-commit: 6039d583686a61426892cd3f3d5f7771837b8d35
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Brice Bartoletti <bib@odoo.com>