[FIX] account_followup: cron task interates over all companies
How to reproduce the bug ?

- install account_accountant
- create a second company
- in the accounting module, in Settings>Follow-up levels, create for
each company, 2 follow-up levels. Each level must have the "Save an
Email" and "Auto Execute"
- for each company, create one invoice with a 1 month old invoice date
- post the invoices
- in debug mode, in the Settings app, go to Technical > Scheduled
Actions and select the "Account Report Followup; Execute followup"
action. click on the "Run manually button".
- Go back in the accounting module and select Customers > Follow-up
Reports.

What is the bug ?

When you create a new company, only the admin has access to it.
Nevertheless, the cron task processing the follow-up reports is executed
by Odoobot who doesn't have access to the new company except if the user
has given access to it.

closes odoo/enterprise#29086

X-original-commit: 55eb5f5e37ef4c37e725d338da64623bc98e40bf
Signed-off-by: Adrien Minet <admi@odoo.com>
Signed-off-by: Minet Adrien (admi) <admi@odoo.com>