[IMP] sale_amazon: handle feeds status
With the Amazon Connector, confirming a delivery sends a so-called
"feed" to Amazon. When the feed is created, Amazon responds to the
request with only the feed id. It is then processed asynchronously by
Amazon and its final status is never communicated by Amazon to Odoo,
hence preventing the user from being aware of synchronization issues.

Keeping the feed id allows us to query Amazon about its status and
process it to know if the synchronization succeeded or not, to keep
track of what happens to the picking, and inform the user that he needs
to change something to his picking before sending the information back
to Amazon should the need arises.

task-2801067

closes odoo/enterprise#28970

Related: odoo/upgrade#4146
Related: odoo/documentation#3193
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>