[FIX] l10n_nl_reports: fix typo in domain->control_domain
Before this commit, there was a small typo with no dramatic impact on the Dutch p&l. Basically, the domain field was modified 2 times, instead of having 1 control_domain and 1 domain. The last domain field was the correct one, so no real impact, but it's still a typo that needs fixing.

closes odoo/enterprise#29173

X-original-commit: 146b73eeaeaa2e186861211f580294ae8d451fe4
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Popeler Antoine (anpp) <anpp@odoo.com>