[FIX] hr_contract_salary: use contract template working hours
Employees created through the salary configurator
flow (/salary_package/simulation/contract/) always ended up with the
default for resource_calendar_id which is
self.env.company.resource_calendar_id. This flow usually occurs with
the public user so this can be wrong when the company on the public
user is not the company the employee will be working at.

To solve this use the resource_calendar_id specified on the contract
template (that's what it's there for anyway). This field is required
in the view so we can rely on it being set.

opw-2896517

closes odoo/enterprise#29111

X-original-commit: a1816d643b159e3c882cdd36722ea1d9ff9f6666
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>