[IMP] project_account_budget: add AA as default AA in budget line
Before this commit, when the user wants to add a budget from the project
update of a project, the budget line added does not have the AA of the
current project by default.

This commit adds the AA of the project by default in the budget line
when the user is in the dialog form view of Budget to add a new budget
for the project.

task-2861207

closes odoo/enterprise#29304

Related: odoo/odoo#95721
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>