[IMP] approvals: Set employee's manager first in approval sequence.
When a sequence is defined on an approval category and the
employee's manager is selected as approver, they were the
last in line to approve. With this commit, they are first in line.

In this commit the sequence field on approver is also fixed, as it
was not saved to the database before.

task-2884137

closes odoo/enterprise#28997

Signed-off-by: Kevin Baptiste <kba@odoo.com>