[FIX] l10n_lu_reports : no balance sheet in xml export
Steps to reproduce:
	- use demo data
	- install l10n_lu_reports
	- Go into accounting>reporting>balance sheet and export to xml
Issue:
	- The fields CA BILAN and CA_COMPP would have no lines

This feature works in 14.0 but not in 15.0.
This is due to a change in the function _get_table() in account.report

Previously it would return the id as a int, but in 15.0 this is a string
containing the full model name followed by a number.

To adapt the current function of 15.0 to the new _get_table() I've
simply applied a regex function to extract the last digit in the id
and cast it to an int. After that, the rest of the function can
keep working in the same way as in 14.0.

closes odoo/enterprise#30020

X-original-commit: ed51fb2f359e1d44330e20994ddf0405ed33bf59
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Vranckx Florian (flvr) <flvr@odoo.com>
Co-authored-by: John Laterre <jol@odoo.com>