[FIX] l10n_es_reports: correct id comparison when generating BOE for mod 347
We used to check whether the generic line id of a grouped by partner line was included in a set of partner ids to know whether or not to inject additional lines into the BOE. This was wrong, as generic line ids and model ids are different concepts ('-res.partner-12' != 12).

Because of that, partners who had already received a value because of the content of the report could receive a second line if they were required in the file (for example if a line had been manually added to the mod 347 BOE wizard for them). This was wrong.

closes odoo/enterprise#29216

X-original-commit: 9e0b4d31ea1eeebcf1fc92838d58a7a0348c5f78
Signed-off-by: Josse Colpaert <jco@odoo.com>