[FIX] l10n_be_reports: 281.50 with currencies
The aim of this commit is to fix the case of bill in foreign
currency.

Previous to this commit:
The amount used for the computation was the amount_total which is potentially
in foreign currency.

After this commit:
The amount used for computation is amount_total_signed which is always in
company currency as required for the 281.50 reporting. (in €)

closes odoo/enterprise#29074

Task: 2820482
Enterprise-pr: https://github.com/odoo/enterprise/pull/28349
X-original-commit: 4e9f1538150d5991c92e73eeb33656dd97d16ed6
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Brice Bartoletti <bib@odoo.com>