[IMP] sale_subscription: standardize validation transaction reference
Before this commit, the reference of a validation transaction followed
this format: “validation-20220222072710”, which did not match the
standard format for refund transactions (”R-S000001”). Moreover, some
providers limit the length of transaction references which forced us to
truncate the quite long reference.

With this commit, future validation transactions will have a reference of
the format: “V-20220222072710”.

task-2830826

closes odoo/enterprise#29076

Related: odoo/odoo#94921
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>