[IMP] knowledge: generate private tutorial article per user
To ease onboarding on the knowledge application, this commit allows to generate
one tutorial/welcome article in the private section of each existing or new
user.

That tutorial article can directly be edited by the users to test all the
functionnality of the Knowledge App.

To avoid generating those articles, a context key can be used:
- knowledge_skip_onboarding_article

Task-2674460

X-original-commit: 0777c9c5c5eaba3e95d35ba52d8b8156f56240c5
Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: David Beguin <dbe@odoo.com>
Co-authored-by: Fabio Barbero <faba@odoo.com>
Co-authored-by: Thibault Delavallee <tde@odoo.com>
Co-authored-by: Aurelien Warnong <awa@odoo.com>