[FIX] knowledge: unarchive multiple articles
Currently, when we unarchive a single article, it works properly by
making it active again and redirecting user to that article. But
when we try to unarchive multiple articles, it throws traceback.

This commit fixes the issue by properly handling the action for
unarchiving, which now will simply reload the same action in
case of multiple articles. Note that behavior for unarchiving
single article is still the same.

taskID-2869659
Original PR odoo/odoo#92851

X-original-commit: 0565b48ffa50d7ef06308f0f8228091adbf1f361
Co-authored-by: Umesh Gupta <umgu@odoo.com>