[IMP] knowledge: prevent article reload on chatter reload
Currently, when the chatter has to reload, it triggers a reload event for the
parent view. In the case of Knowledge, we want to be able to exclusively reload
the chatter because nothing changed in the article.

In Knowledge, if the param `keepChanges` is used with reload, the full form view
does not need to be fully reloaded. In this case, only the chatter will be
updated.

Task-2858428
Original PR odoo/odoo#91969

X-original-commit: da1206f1c8bbff90dd2a613eca1307bc2a9712cd
Co-authored-by: Damien Abeloos <abd@odoo.com>