[IMP] knowledge: restyle the chatter button and size
When the chatter is in side mode (lg+ sizes), its size can become quite small
(since it is proportional with the width of the viewport). Add a static minimal
width so that the display stays readable.

Change the border color of the buttons in the knowledge header so that when the
chatter is displayed, it has a better contrast (the chatter button is a toggle
on-off).

Task-2858428
Original PR odoo/odoo#91969

X-original-commit: af2bff150bdfb50a013258c7fd9f8dffc2c0516a
Co-authored-by: Damien Abeloos <abd@odoo.com>