[IMP] knowledge: add no content helpers and sample data
PURPOSE:

- Add no content helpers in the favorite and member views, as well as
on the main "article" view: when no article exists, a smiling face
suggests to create one, when no favorite relation exists, a smiling
face suggests to add an article to the favorites and explains how to
do it, and when no member relation exists, the purpose of creating
one is explained.

- Add samples in background when there are no records in the members
and favorite views.

- Set the favorite list views as editable=bottom to allow
modifications from the list view.

- Remove the sequence handle in the favorite list view to prevent
reordering, because it led to strange behaviors since the favorite
articles are ordered by user id.

- Disallow users to create or edit records in the members view, as
creation and edition are normally handled by custom methods, that won't
be triggered when creating or editing from this view.

- Allow users to search for articles based on their content.

- Make the favorite and member views open to admin users in debug mode
only, instead of all internal users in debug mode.

These additions have been made with the aim of improving the user
experience.

Task-2852946
Original PR odoo/odoo#93690

X-original-commit: 3e1e31d5f05ceb1cc0da69971989a62f3f13aa8a
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Aurelien Warnon <awa@odoo.com>