[FIX] knowledge: fix resequencing articles
Purpose
=======

Properly display and resequence articles based on sequence field. Articles are
currently not sorted based on sequence, and when moving articles outside of a
parent resequencing is not applied.

Specifications
==============

Properly resequence articles not only if the parent_id value is non-false but
also if it's in the vals and set to False (when moving an article to the root).

Task-2883368
Original PR odoo/odoo#93597

X-original-commit: 58e48767c2c63fd03e92442152b46d2b2db2bf31
Co-authored-by: Fabio Barbero <faba@odoo.com>