[FIX] knowledge: fix member inheritance in share panel
Description of the issue:
=========================

Suppose we have 3 articles with the following structure:

root
   L child
         L grand child

When granting the "write" permission to a user in "root", this user
will inherit this permission in "child" and in "grand child".
When we then restrict this user with the "no access" permission in
child, this permission is inherited by grand-child (correct).
Then, when a member is invited in the root article, the right is not
herited by "child" nor "grand child" (correct). However, this
permission is shown as being inherited from "root" in the permission
panel in "grand child" (incorrect: "grand child" does not have this
permission, and is in fact inherited from "child" as expected).

Solution:
=========

Add the boolean "article_is_desynchronized" check in the inner
join condition of the recursive SQL request used to fetch the
permissions of an article and its parents.
This boolean check allows to prune the recursion when an article
is desynchronized, so that the permissions of the parents of this
desynchronized article are not fetched, since they are not
inherited in the desynchronized articles nor in its children.

Some tests have been added to ensure the correct behaviour of the
share panel. Especially, the scenario described in the description of
the issue is tested.

Task-2866950
Original PR odoo/odoo#93215

X-original-commit: 97aa5d8d717dd1bee66183cb3f47402db38b5411
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Aurélien Warnon <awa@odoo.com>