[FIX] industry_fsm_sale: execute a test in post_install
Before this commit, all tests in the TestFsmFlowSale class are launched
during the installation of the module. However, one of those tests
creates some invoices and so a journal has to be defined in the company.

This commit moves the others tests in new class to continue to launch
thoses during the installation of the module and the change the
tests execution of TestFsmFlowSale (containing only the `test_fsm_flow`
method in which an invoice is created) to launch this test after
installing all modules.

closes odoo/enterprise#29208

X-original-commit: 649d9f6d8c4c4aeef3c997c5d63bbe8986e1fe86
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Xavier <xbo@odoo.com>