[IMP] approvals: convert product lines uom, description onchange to compute
Convert the onchange on `product_id` to compute fields

This allows to create a `approval.product.line` record without
the need to specify the description or uom during the creation nor calling the onchange
method. The description and uom defaults from the product.

This way, if the database is not configured to use multi uom,
the users do not have the specify the UOM to use.
For instance during an XMLRPC call to create an approval request with a
product line.

closes odoo/enterprise#29175

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>