[FIX] sale_subscription_dashboard: traceback on KPI filters
When merging sale_subscription to sale_order models, the sales team filter
in sale_subscription_dashboard started throwing a traceback due to a wrong
naming of a table. Also, the categories filter was not working properly
due to a rename of the account_tags relational table. This commit fixes both
issues.

X-original-commit: 55c1014e4a08758a0af77d7e93ed2abecea97f1c
Part-of: odoo/enterprise#29705