[FIX] payment_sepa_direct_debit: read the provider from the record
Due to an oversight, the "Save my payment details" checkbox was shown on
the inline payment form of SEPA Direct Debit acquirers, which should
never happen because the transaction is *always* tokenized with those.

With this commit, the `_is_tokenization_required` method is slightly
refactored to read the provider from the current `payment.acquirer`
record rather than from the kwargs. This conveniently fixes the issue
and prevent it from happening again elsewhere.

closes odoo/enterprise#29237

X-original-commit: c48e22ca0954d027c46d4fe6639e8582cb8c3b21
Related: odoo/odoo#95519
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>