[IMP] {website_}appointment: add searchBar suggestions
Purpose
=======
Use the new website searchbar which displays suggestions

Specifications
==============
Converting some instance methods of the controller to be
static in order to reuse the logic for the appointments
selection while computing the search results.

This new search bar respects the URL-provided filters
and a user with a valid invite token is able to search
for unpublished appointment types.

Task-2888646

closes odoo/enterprise#29219

Related: odoo/upgrade#3665
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>