[FIX] industry_fsm: use `float_compare` to compare float fields
Before this commit, the test to compare the values in Fsm Task Analysis
with the values stored in the task sometimes fails because a float value
could have more than 2 decimals and so it will consider as is not equal
to the other value.

This commit replaces the `assertDict` by `assertEqual` for each key value
in the both dictionary and in that assertEqual the `float_compare` is
used to compare the both value until the 2 first decimals.

closes odoo/enterprise#29224

X-original-commit: 1245fe41670f7
Signed-off-by: Kevin Baptiste <kba@odoo.com>