[MOV] knowledge: extract method to detach children
This commit extracts the logic in "action_archive" that detaches children
articles to which the current user does not have access to into a separate
method. See 'KnowledgeArticle._detach_unwritable_descendants()' for details.

This is a preliminary work to introduce the next fixes that will apply that
logic when making an article private as well. There should be no functional
change with this commit.

Task-2859616
Original PR odoo/odoo#94836

X-original-commit: 1f806437984240771f9a59c0255c536ff9fa1bf4
Co-authored-by: Aurélien Warnon <awa@odoo.com>