[FIX][IMP] knowledge: lessen usage of member writable constraint
Purpose of this commit is to have less check on "is writable" constraint from
member model. With a new context key we can now bypass the constraint from
member model when we are sure it is globally verified. As in some cases we
modify members from an article, the constraint will be checked at article
level with all members being updated. In some cases doing specific checks
might lead to issues when the constraint is evaluated in a transient state
e.g. when a member is removed while new write members are not yet added.

Task-2859616
Original PR odoo/odoo#94836

X-original-commit: 9e482d8e730edbbd8d20a3e4167409e4bf0985e3
Co-authored-by: Aurélien Warnon <awa@odoo.com>
Co-authored-by: Thibault Delavallee <tde@odoo.com>