[FIX] knowledge: correctly handle writer check when removing memberships
This commit fixes the check that verifies that the article always has a member
with write access on the article.

Indeed, the current code was using a database request to check for members with
write access in a "api.constrains" check, where the data being updated is not
flushed into database yet.

It now correctly uses fields access directly, as stated by the method docstring.
For convenience, we extracted the "_has_writer_member" method to be re-usable.

An additional test case has been added to enforce the behavior.

Task-2859616
Original PR odoo/odoo#94836

X-original-commit: 4e9403a275e8d53a3121ff8ab22c30179cc8833f
Co-authored-by: Aurélien Warnon <awa@odoo.com>