[FIX] knowledge: correctly handle desync status when detaching children
This commit fixes the case when we have desynchronized children for which the
current user does not have write access and we move the parent to the private
section (or archive the parent).

In that case, children that are desynchronized should not have the members from
their parents copied onto them when detaching.

See '_detach_unwritable_descendants' and associated tests for more details.

Task-2859616
Original PR odoo/odoo#94836

X-original-commit: 3ab27cdaaec24372f2857617b42a944f0212c6b3
Co-authored-by: Aurélien Warnon <awa@odoo.com>