[FIX] knowledge: archive children with current user
Surprisingly ``_filter_access_rules_python`` returns a sudo-ed recordset. We
do not want to call super in a write as sudo, even if the recordset has been
filtered based on current user's ACLs.

In this commit we reset the recordset environment to the current user's one.

Task-2859616
Original PR odoo/odoo#94836

X-original-commit: 39bc819836fed3a8bb1d855ec84e2cdb179128e3
Co-authored-by: Aurélien Warnon <awa@odoo.com>