[FIX] project_enterprise_hr: change create_date during creation
Before this commit, the `create_date` field of the employee used in the
unit test is updated via a sql query after creating that employee. By
doing that, the cache is not invalidated and so the `create_date` still
remains the current date and so the schedule will not based to the
working schedule of that employee since it will be considered as not
working to the company before the current date, and so the working
schedule of the company will be used instead.

This commit updates the `create_date` directly during the creation of
that employee and avoid doing a sql query for that. By doing this, the
employee will be considered as working on the company since the
`create_date` is before the dates to plan the tasks and so the working
schedule will not be used.

closes odoo/enterprise#29244

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>