[FIX] account_intrastat: unknown country code QV
The unknown country code used for the country of origin of the goods is
incorrect. QU refers to an unknown country in the general context, in
the context of intrastat (intra-EU trade) the correct country code is
QV (an unknown EU country).

This commit changes the QU code to QV inside the query.

In 13.0 the code is already QV, however the line is refactored to match
the code of the later stable branches. Replacing the conditional with a
'COALESCE'

The intrastat_product_origin_country_name is also adjusted in this way

This distinction is present both in the Dutch specifications:
https://www.cbs.nl/-/media/cbsvooruwbedrijf/international-trade-in-goods/manual_statistical_declaration_itg_2022.pdf
and in the Belgian specifications:
https://www.nbb.be/doc/dq/e_pdf_ex/geonomenclatuur2013en.pdf
and is generalised in the EU's genomenclature reporting:
https://ec.europa.eu/eurostat/documents/3859598/5889816/KS-BM-05-002-EN.PDF.pdf/62e82e02-4632-438b-a284-e8fe75984b32?t=1414781146000

closes odoo/enterprise#29301

X-original-commit: eb216ba0b55afa306cdcbf178dcf123a33790769
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>
Signed-off-by: Brice Bartoletti <bib@odoo.com>