[FIX] knowledge: _search method should not use self when calling model methods
This commit makes _search methods on fields use the model instead of 'self'
when calling model permissions methods (that use self to filter out their
output result).

Those _search methods are always called when searching knowledge.articles as
their related fields are part of ACLs definitions.

Globally speaking, calling 'search' on self should *never* filter out records
based on what is currently contained in the self recordset.

Technical note:

The difference in the move_to query counter is due to the fact that we now
correctly resequence articles after they are moved.

Indeed, before that, the "search" in the "_resequence" method was being limited
to fewer articles than intended, because of what is explained here above.

Task-2852909

X-original-commit: 396b44e2bec62ddcd0b2a3b974de8a62242c4b76
Part-of: odoo/enterprise#29611