[FIX] knowledge: filter out non accessible article in search and move to
Before this commit:
As admin can see all the article even if they have no explicit access to them,
admin will find all the private article of other users in the "move to"
proposal dropdown, in the search results (via the command palette) and in
the results in /article command.

After this commit:
The result of the "move to dropdown" is filtered to list only the article the
user has access to. Same for the search results via the command palette and same
for the results in /article command.
The admin can still find the other's private article using the advanced search.

Task-2852909

X-original-commit: 64dbfb31b8b1a76ab4e5df16f5223a57ca89c654
Part-of: odoo/enterprise#29611