[IMP] (website_)helpdesk(_knowledge): helpdesk-knowledge integration
PURPOSE

This task aims to encourage customers to go through Knowledge articles
in the Helpdesk's help center.
Hopefully, they find an answer to their issue by themselves
instead of raising a ticket. This would help reduce the support workload
and allow customers to get a solution without delay.

SPECIFICATIONS

helpdesk.team form view > add this option:
        > name: Knowledge
        > description:
Centralize, manage, share and grow your knowledge library
        > disabled by default

If the 'knowledge' and the 'website form'
features are enabled, display the following on the Help webpage of the team:
        > fa-book icon
        > Articles
        > Search our documentation for answers to common questions
        > 'Browse Articles' button that should redirect the current user
to the knowledge articles he can access.

Note:
Added one to the query count of the following tests
in /knowledge:TestPerformanceCase:
	> test_article_move_to()

As /website_helpdesk_knowledge:KnowledgeArticle._check_helpdesk_constrains()
checks that, when parent_id is set or is_published is set to False,
this article is the website_article_id of a HelpdeskTeam.
It true, it raises a ValidationError.

related: https://github.com/odoo/odoo/pull/97739
task-2890043

closes odoo/enterprise#29283

Related: odoo/odoo#97739
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>