[FIX] web_studio: adapt studio approval component to OWL
The components used to handle the studio approval feature (that is:
ask the server if one can execute an action, when the button has been
marked with studio_approval="True" attribute) have been adapted to OWL.
All the data logic is handled by a small class that a hook can retrieve
in order to give a component the approval features.

Potentially all views can now support studio_approval="True" on buttons
since only the relevant component (ViewButton) and compiler
(ViewCompiler) have been patched to handle the feature.

Part-of: odoo/enterprise#23069