[FIX] l10n_us_payment_nacha: right align the immediate fields
In cases where the nacha_immediate_* fields don't consist of 10
characters we should pad with spaces and align to the *right*.

It's a bit more logical and this is hinted to in the Bank of America
spec [1]:

bNNNNNNNNN | Immediate Destination | Number that identifies site.

where “b” indicates a blank space.

[1] p9 (p13 in PDF) https://files.nc.gov/ncosc/documents/eCommerce/bank_of_america_nacha_file_specs.pdf

opw-2858885
opw-2896315

X-original-commit: d31fe41af31925d8913fd20530e5bdf321db27b8
Part-of: odoo/enterprise#29779