[FIX] l10n_us_payment_nacha: calculate the right block count
The block count represents how many blocks are in the file. It's
calculated by taking the amount of records (3 per payment + header +
footer) and dividing it by the blocking factor (always 10). Some
documentation recommends to floor or ceil this result, but a more
robust way is to pad the file with extra records (999...) until there
are a multiple of 10 records. This avoids the whole rounding problem.

Bank documentation is not very clear on this, but the ACH Universal
documentation has a good explanation [1].

[1] https://www.treasurysoftware.com/ach/how-do-i-create-an-ach-nacha-file-detailed-file-format.aspx

opw-2858885
opw-2896315

X-original-commit: 2be3fb51d0c0c905734f131931ee1f0fde15d908
Part-of: odoo/enterprise#29779