[IMP] knowledge: add Table of Content command
Allow users to create a table of contents dynamically listing all
headers in the page. When clicking on a section name, the page will
scroll (both in back end and front end view) to the selected heather and
highlight it for a brief interval of time.

This table of content will automatically update once the user starts modifying
the headers structure of his article.

A few tests were added to ensure the "layout" result we want based on available
headers within the article body.

Additionally, the "HR FAQ" data was updated to clean the header structure and
add a table of content to showcase the new feature.

Task-2818474

closes odoo/enterprise#29303

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Co-authored-by: Aurélien Warnon <awa@odoo.com>
Co-authored-by: julien-banken <jbn@odoo.com>
Co-authored-by: abd-msyukyu-odoo <abd@odoo.com>
Co-authored-by: fbarbe00 <faba@odoo.com>