[IMP] knowledge: improve article tree usability
Prevent small movements to trigger an article move:

- Change the move tolerance of the article tree. Now, the placeholder
  indicating the destination position will move only if the dragged item hovers
  at least 50% of the height of the previous/next item (instead of 1px before).

- Add a delay and a minimal moving distance before starting to drag an article.
  This will help to avoid unwillingly dragging an article when clicking on it.

The moveArticle function in the main flow tour is updated to reflect this
change. A new step is added after the move of an article to wait for completion
before continuing.

Reduce the move scroll speed in the article tree. When dragging an item below
the view, the reduced speed allows more control over the destination.

Force close the emoji picker when dragging an article.

Prevent dragging an article by clicking on an emoji in the emoji picker.

Add :hover and :active colors in the article tree to better reflect what the
user can click on.

Unhide ui-sortable-handle in Knowledge on small screens, so that the article
tree does not become unusable. Drag and drop does not work on mobile, but
switching article by clicking (touching) on them does, so it makes sense to be
able to see it.

Move the emoji picker outside of the article tree. Only use one emoji picker for
the entire form view instead of one per article (which could be a lot). The
previous location was also an issue for the article handle (it was possible to
drag&drop an article from its emoji picker).

Add a transparent div taking the entire viewport when displaying the emoji
picker. This will force the user to close the emoji picker before continuing to
use the view.

Task-2883290

Part-of: odoo/enterprise#29315