[FIX] spreadsheet_edition,timesheet_grid: properly update internal state
Since last owl update, the way t-att-value works has been slightly
changed, and now, each owl update will force the input to match the
internal state of owl, even if it has been modified in the meantime.

Because of that, the timer in the timesheet grid header could behave
in a wrong way: if the user would modify the input state manually, then
the internal state does not match the input, and any subsequent render
would override the user value.

We fix this by making sure that the internal state is also updated.

Some issue appears in model_selector and spreadsheet_name. This commit
fixes them by properly managing the state of the input value.

closes odoo/enterprise#29350

Related: odoo/odoo#95804
Signed-off-by: Samuel Degueldre <sad@odoo.com>
Signed-off-by: Géry Debongnie <ged@odoo.com>