[FIX] documents: fix crash upon using deleted tags
Before this commit:
Install documents

In the aliases, in the alias_defaults, change the default tag_id, folder_id or create_share_id to one that does not exist (for example, for the default alias inbox, change from
  {'tag_ids': [(6, 0, [1])], 'folder_id': 1, 'partner_id': False, 'create_share_id': 2}
  to  {'tag_ids': [(6, 0, [1000])], 'folder_id': 1, 'partner_id': False, 'create_share_id': 2})

Then, send a document to that alias.

It will not find the relevant tag_id, folder_id or create_share_id and will send back an error without registering the document.

After this commit:

If a tag_id does not exist, remove it (documents.tag are not mandatory)
Other errors are not handled by this since they may not be erronous
without manual modification.

OPW-2786973

closes odoo/enterprise#29418

X-original-commit: 976b336870bd5ad63eb1e4f119aedb5e9b55daf1
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>