[PERF] knowledge: introduce a query counter test for the tree panel
This commit simply introduces a query counter test for the tree panel knowledge
route. This will allow us to optimize that route in following commits.

While we are at it, we also add a quick query counter test for the
'action_home_page' method, as it is called every time the main Knowledge menu
is clicked.

Task-2791000

closes odoo/enterprise#30745

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>