[IMP] knowledge: allow article drop in the shared section
Allows to drop an article under a shared article (it will inherit the shared rights
of the shared parent), and to resequence a shared root article.
Previously, dropping an article in the shared section was prevented because
if the dropping destination is at the root of the shared section, the access rights
can not be inferred. But every other move is now allowed with this commit.

Replace "is_private" parameter in knowledge.article.move_to method by
"category", since the previous signature did not allow to identify whether an
article was moved at the end of the workspace section or at the end of the
shared section.

In move_to, infer the category from parent_id or before_article_id or the
provided category in that order to avoid inconsistent requests. Prevent moving
an article as a root of the shared section if it was not already one. Those
checks increase the query count of
KnowledgePerformanceCase.test_article_move_to.

Task-2867052

X-original-commit: 4238df0045ea587d7e0eda17a920158bde6cf9dd
Part-of: odoo/enterprise#30312