[PERF] knowledge: introduce '_parent_store' to speed up hierarchy requests
Currently, searching through the knowledge.article hierarchy by either going up
using the 'parent_of' operator or going down using the 'child_of' operator is
sub-optimized because the ORM needs to recursively go up or down the hierarchy
using the parent_id/child_ids fields.

This commit enables the _parent_store/parent_path on the model, allowing the
ORM to store the path to the article parent in a separate field.
For example, if the article ID 8 has a parent ID 4 which itself has a parent
ID 2, the stored parent_path or article 8 will look like: '2/4/8/'.

This notably speeds up the side tree panel display by saving up a few queries.

It has the small drawback of:
- Needing an extra field/index
- Having to update the hierarchy when it's changed
  As the hierarchy update is done AFTER creation, this adds one query to
  performance tests that create articles.
  (see '_parent_store_create' for details)

However, as we will not have a lot of article create/write, and probably a LOT
more reads, it seems more profitable to optimize the search querying.

Technical side-note: As the '_parent_store_update' method also checks for
potential recursions, tests in the knowledge module have been adapted.
They are now split as the transaction is rolled back after the UserError, which
caused CacheMiss when the tests were merged together.

Task-2791000

Part-of: odoo/enterprise#29442