[FIX] knowledge: add default file icon on emoji removal
When removing the emoji of an article, the emoji is removed from the
article display name in the sidebar, which makes it difficult to select
a new one (small button showed on hover). When reloading or selecting
another article, the default file icon reappears.

Now, when removing an emoji, the emoji is directly set to the default
file icon in the sidebar to avoid having this small hidden button.

Task-2905851

closes odoo/enterprise#29395

X-original-commit: 8813422d3007d28291848e887b21a27280aa83d4
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>