[FIX] knowledge: fix access error with share panel
When an article inherits its members from its parent and the user does
not have "read" access on the parent article, the user gets an access
error when opening the share panel of the child article.

To reproduce the issue:

As "Mitchell Admin" (admin user):
1. Create a new article in workspace
2. Invite any user on the article
3. Create a child article
4. Set the internal permission of the parent article to "none"
5. Publish the child article
6. Copy the url of the child article

As "Marc Demo" (demo user):
1. Access the article from the url you copied at step 6
2. Open the permission panel

The `get_article_permission_panel_data` function does not check that the
user can have access to the parent article before accessing its properties.
As a result, the user can get an access error when opening the share panel.
To fix the issue, we will do a sudo call on the parent article before
accessing its properties. The user will then no longer get an access error.

task-2893614

X-original-commit: ca82d88a41ffbaeeb82ce99c7b6cbfcc9a38757c
Part-of: odoo/enterprise#30430