[PERF] knowledge: store the article active status on linked favorites
Currently, every time we fetch knowledge.article.favorite records, we force the
domain to only retrieve the ones for active articles.

This causes the ORM to evaluate the ACLs for our knowledge.article model,
creating extra (and rather costly) queries.

To avoid this, we now store the 'is_article_active' field on the
knowledge.article.favorite model.

This increases the amount of 'write' operations (to sync the fields) but should
not be very problematic as it's only done when archiving/unarchiving articles.

This change will be notably beneficial to:
- The loading of the side tree panel
- The 'action_home_page' method
  Which is called every time the user clicks on the Knowledge main menu

Task-2791000

Part-of: odoo/enterprise#29442